cleanup (ContentCard): move to `~/components/Base`
