+++ pages/index.vue
@@ -82,8 +82,8 @@ import { analytics } from '~/lib/firebase'
 import TopAlert from '~/components/TopAlert'
 import PopupCampaign from '~/components/PopupCampaign'
 import Section from '~/components/Base/Section'
-import ContentCard from '~/components/ContentCard'
-import ContentCardButton from '~/components/ContentCard/ContentCardButton'
+import ContentCard from '~/components/Base/ContentCard'
+import ContentCardButton from '~/components/Base/ContentCard/ContentCardButton'
 import BannerCarousel from '~/components/Homepage/BannerCarousel'
 import EventStatistics from '~/components/Homepage/EventStatistics'
 import RecentNewsCarousel from '~/components/Homepage/RecentNewsCarousel'
